# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PGATS-02 API",
    "version": "1.0.0",
    "description": "API em memória para aprendizado de testes e automação."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Registrar usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "favored": { "type": "boolean" },
                  "balance": { "type": "number" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado" },
          "400": { "description": "Dados inválidos" },
          "409": { "description": "Usuário já existe" }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Listar usuários (sem senhas)",
        "responses": { "200": { "description": "Lista de usuários" } }
      }
    },
    "/transfers": {
      "post": {
        "summary": "Criar transferência",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fromId": { "type": "string" },
                  "toId": { "type": "string" },
                  "amount": { "type": "number" }
                },
                "required": ["fromId", "toId", "amount"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transferência criada" },
          "400": { "description": "Dados inválidos ou fundos insuficientes" },
          "403": { "description": "Regra de favorecido violada" },
          "404": { "description": "Usuário de origem ou destino não encontrado" }
        }
      },
      "get": { "summary": "Listar transferências", "responses": { "200": { "description": "Lista" } } }
    }
  }
}